// src/app/model/BookingSpa.ts
//create data field in database
export interface BookingSpa 
{
    id?: any;
    customerName: string;
    contactNumber: string;
    catName: string;
    remark: string;
    checkInDate: string;
    checkOutDate: string;
    timeIn: string;
    timeOut: string;
    // catFood: Selection;
    // createdAt: any;
}